PublishCapableRepository tweaks
