+++ src/main/java/org/craftercms/studio/api/v2/repository/GitPublishCapableRepository.java
@@ -18,7 +18,7 @@ package org.craftercms.studio.api.v2.repository;
 
 import org.craftercms.studio.api.v1.exception.ServiceLayerException;
 import org.craftercms.studio.api.v2.dal.publish.PublishPackage;
-import org.springframework.util.ObjectUtils;
+import org.craftercms.studio.api.v2.repository.publish.GitPublishChangeSet;
 
 import java.io.IOException;
 import java.util.Collection;
@@ -27,20 +27,7 @@ import java.util.List;
 /**
  * Interface for publish operations of a git repository
  */
-public interface GitPublishCapableRepository extends GitContentRepository {
-	/**
-	 * Publishes the given items to the given target
-	 *
-	 * @param publishPackage   the publish package
-	 * @param publishingTarget the publishing target
-	 * @param publishItems     the items to publish
-	 * @param <T>              the type of the {@link PublishItemTO} objects
-	 * @return the change set listing the affected paths and new commit id
-	 * @throws ServiceLayerException if there is any error while publishing or publishItems is null or empty
-	 */
-	<T extends PublishItemTO> GitPublishChangeSet<T> publish(PublishPackage publishPackage,
-								 String publishingTarget,
-								 Collection<T> publishItems) throws ServiceLayerException, IOException;
+public interface GitPublishCapableRepository extends GitContentRepository, PublishCapableRepository {
 
 	/**
 	 * Publishes all changes for the given site and target
@@ -63,33 +50,4 @@ public interface GitPublishCapableRepository extends GitContentRepository {
 	String initialPublish(PublishPackage publishPackage, List<String> ignorePaths,
 			      String target) throws ServiceLayerException;
 
-	/**
-	 * Store the result of a publish operation
-	 *
-	 * @param successfulItems the paths that were updated
-	 * @param failedItems     the paths that failed to publish, mapped to the error message
-	 * @param <T>             the actual type of the {@link PublishItemTO} objects
-	 */
-	record GitPublishChangeSet<T extends PublishItemTO>(String commitId,
-							    Collection<T> successfulItems,
-							    Collection<T> failedItems) {
-
-		/**
-		 * Check if there are successfully published changes
-		 *
-		 * @return true if the package had successful changes and a commit was created, false otherwise
-		 */
-		public boolean completed() {
-			return !ObjectUtils.isEmpty(commitId);
-		}
-
-		/**
-		 * Check if there are failed items
-		 *
-		 * @return true if failed items list contains items, false otherwise
-		 */
-		public boolean hasFailedItems() {
-			return !ObjectUtils.isEmpty(failedItems);
-		}
-	}
 }
