+++ src/main/java/org/craftercms/studio/api/v2/repository/PublishCapableRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2024 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -18,26 +18,15 @@ package org.craftercms.studio.api.v2.repository;
 
 import org.craftercms.studio.api.v1.exception.ServiceLayerException;
 import org.craftercms.studio.api.v2.dal.publish.PublishPackage;
-import org.craftercms.studio.api.v2.repository.GitPublishCapableRepository.GitPublishChangeSet;
-import org.springframework.util.ObjectUtils;
+import org.craftercms.studio.api.v2.repository.publish.GitPublishChangeSet;
 
 import java.io.IOException;
 import java.util.Collection;
-import java.util.Map;
 
 /**
  * Interface for content repositories that support publishing
  */
 public interface PublishCapableRepository {
-	/**
-	 * Execute initial publish for given site
-	 *
-	 * @param publishPackage the package to publish
-	 * @return commit id of the initial publish.
-	 * After this method runs, the returned value is the last
-	 * commit in the published repository for the target branch
-	 */
-	InitialPublishChangeSet initialPublish(PublishPackage publishPackage, String target) throws ServiceLayerException;
 
 	/**
 	 * Publishes the given items to the given target
@@ -50,17 +39,6 @@ public interface PublishCapableRepository {
 	 * @throws ServiceLayerException if there is any error while publishing or publishItems is null or empty
 	 */
 	<T extends PublishItemTO> GitPublishChangeSet<T> publish(PublishPackage publishPackage,
-								 String publishingTarget,
-								 Collection<T> publishItems) throws ServiceLayerException, IOException;
-
-	record InitialPublishChangeSet(String commitId, Map<String, Integer> failedItems) {
-		/**
-		 * Check if there are failed items
-		 *
-		 * @return true if failed items list contains items, false otherwise
-		 */
-		public boolean hasFailedItems() {
-			return !ObjectUtils.isEmpty(failedItems);
-		}
-	}
+															 String publishingTarget,
+															 Collection<T> publishItems) throws ServiceLayerException, IOException;
 }
