+++ src/main/java/org/craftercms/studio/api/v2/repository/blob/StudioBlobAwareContentRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2024 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -16,9 +16,12 @@
 package org.craftercms.studio.api.v2.repository.blob;
 
 import org.craftercms.studio.api.v1.exception.ServiceLayerException;
+import org.craftercms.studio.api.v2.dal.publish.PublishPackage;
 import org.craftercms.studio.api.v2.repository.ContentRepository;
 import org.craftercms.studio.api.v2.repository.GitContentRepository;
 import org.craftercms.studio.api.v2.repository.PublishCapableRepository;
+import org.craftercms.studio.api.v2.repository.PublishItemTO;
+import org.craftercms.studio.api.v2.repository.publish.GitPublishChangeSet;
 
 /**
  * {@link ContentRepository} extension that provides blob aware operations.
@@ -33,4 +36,14 @@ public interface StudioBlobAwareContentRepository extends GitContentRepository, 
 	 * @throws ServiceLayerException if an error occurs during the operation
 	 */
 	void duplicateBlobs(String sourceSiteId, String siteId) throws ServiceLayerException;
+
+	/**
+	 * Execute initial publish for given site
+	 *
+	 * @param publishPackage the package to publish
+	 * @return commit id of the initial publish.
+	 * After this method runs, the returned value is the last
+	 * commit in the published repository for the target branch
+	 */
+	GitPublishChangeSet<? extends PublishItemTO> initialPublish(PublishPackage publishPackage, String target) throws ServiceLayerException;
 }
